requirements.txt
# Smart Sensor Data Dashboard - Full Project Dependencies
# This file contains all Python dependencies required for the complete project
# Note: The demo script (demo/run_demo.py) requires only Python 3.8+ standard library
# 
# Core Web Framework
fastapi==0.115.0
uvicorn==0.31.0

# Data Processing & Analysis
pandas==2.2.2
numpy==1.26.4

# Database
# sqlite3 is built into Python 3.x, listed for clarity
# PostgreSQL support
psycopg2-binary==2.9.9

# Template Engine
jinja2==3.1.4

# HTTP Client & API Testing
requests==2.31.0
httpx==0.27.0

# Data Validation & Serialization
pydantic==2.6.3

# Environment Configuration
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.1.0

# Development & Testing
pytest==8.0.2
pytest-asyncio==0.23.5
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Security
cryptography==42.0.5

# Optional: Performance & Caching
redis==5.0.1
celery==5.3.4

# Optional: Monitoring & Metrics
prometheus-client==0.20.0

# Optional: Email Notifications
smtplib  # Built into Python standard library

# Optional: Data Visualization (if needed for additional charts)
matplotlib==3.8.3
seaborn==0.13.2

# Optional: Machine Learning (for advanced analytics)
scikit-learn==1.4.1

# Optional: Time Series Analysis
statsmodels==0.14.1

# Optional: Database ORM (if migrating from SQLite)
sqlalchemy==2.0.27

# Optional: Background Tasks
apscheduler==3.10.4

# Optional: WebSocket Support (for real-time updates)
websockets==12.0

# Optional: File Processing
openpyxl==3.1.2  # Excel file support
xlrd==2.0.1      # Legacy Excel support

# Optional: Image Processing (for reports)
pillow==10.2.0

# Optional: PDF Generation (for reports)
reportlab==4.1.0

# Optional: Configuration Management
pyyaml==6.0.1

# Optional: Date/Time Handling
python-dateutil==2.8.2
pytz==2024.1

# Optional: Data Export/Import
tabulate==0.9.0

# Optional: API Documentation
fastapi-openapi==0.1.0

# Optional: CORS Support (already included in FastAPI)
# fastapi-cors==0.0.6

# Optional: Authentication (if needed)
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Optional: Rate Limiting
slowapi==0.1.9

# Optional: Health Checks
healthcheck==1.3.5

# Optional: Metrics Collection
psutil==5.9.8

# Optional: Environment Detection
platform  # Built into Python standard library
os        # Built into Python standard library
sys       # Built into Python standard library
logging   # Built into Python standard library
datetime  # Built into Python standard library
json      # Built into Python standard library
csv       # Built into Python standard library
sqlite3   # Built into Python standard library
http.server  # Built into Python standard library (for demo)
socketserver # Built into Python standard library (for demo) 